+++ packages/eslint-config/next.cjs
@@ -1,0 +1,8 @@
+module.exports = {
+  extends: ["./base.cjs"],
+  env: {
+    browser: true,
+    node: true
+  },
+  rules: {}
+};